Declare the last four operator new overloads with size_t
port/build-port.cmd has not compiled since 010b269ef (2026-09-14). 32-bit MSVC
rejects include/fBase_c.h:158 with

  error C2821: first formal parameter to 'operator new' must be 'size_t'

The rule is on the TYPE, not the width -- all three candidate spellings are four
bytes on this target. mwccarm rejects `operator new(unsigned int)` as an illegal
operator declaration; MSVC rejects `operator new(unsigned long)`. include/types.h
already carries the resolution: a conditional size_t, <stddef.h> under _MSC_VER
and `typedef unsigned long` otherwise, added for exactly this. Most actor headers
already declare `operator new(size_t)`; these four were never converted and still
name the literal type.

No ROM byte moves, because size_t IS unsigned long for mwccarm. Measured with a
build that recompiles every translation unit reaching these headers:

  rombuild -j16 --no-rom   1963 compiled, mismatching 0,
                           module fidelity 106/106 exact, ROM-build PASS

This does NOT make the port green. It clears the compile error, after which four
of fourteen smoke targets build and pass and nine fail to LINK for an unrelated
reason: _ZN5ModelC1Ev and _ZN5ModelC2Ev both define Model::Model() once MSVC
mangles them, and the port's bridges reference ModelBase::Destructor0/Destructor1,
which nothing defines. Those are C++-migration seams in port/hal and need their
own change. Note that ctest reports them as failing tests, but they are "Not Run":
the executables never linked.

+++ include/daBar_c.h
@@ -24,9 +24,9 @@ struct daBar_c : dActor_c {
     virtual s32 Render();
     virtual void OnPendingDestroy();
 
-    /* Leaf adapter until fBase_c::operator new(unsigned long) lands.
+    /* Leaf adapter until fBase_c::operator new(size_t) lands.
        `return new daBar_c()` then routes through the retail allocator. */
-    static void *operator new(unsigned long size) {
+    static void *operator new(size_t size) {
         return _ZN7fBase_cnwEj((unsigned)size);
     }
